Перенес тесты MovieQuizViewControllerMock и StubNetworkClient в отдельные файлы, Поправил строки кода и убрал лишние.
