+++ MovieQuiz/Presentation/MovieQuizPresenter.swift
@@ -1,7 +1,7 @@
 import UIKit
 
 final class MovieQuizPresenter: QuestionFactoryDelegate {
-   
+    
     // MARK: - Private Properties
     
     private var questionFactory: QuestionFactory!
@@ -12,7 +12,7 @@ final class MovieQuizPresenter: QuestionFactoryDelegate {
     private var currentQuestionIndex = 0
     private var correctAnswers = 0
     private let questionsAmount: Int = 10
-
+    
     // MARK: - Initialization
     
     init(viewController: MovieQuizViewControllerProtocol) {
@@ -26,18 +26,18 @@ final class MovieQuizPresenter: QuestionFactoryDelegate {
     // MARK: - QuestionFactoryDelegate
     
     func didReceiveNextQuestion(question: QuizQuestion?) {
-
+        
         guard let question = question else {
             return
         }
-
+        
         currentQuestion = question
         let viewModel = convert(model: question)
         DispatchQueue.main.async { [weak self] in
             self?.viewController?.showCurrentQuestion(step: viewModel)
         }
     }
-
+    
     func didLoadDataFromServer() {
         viewController?.hideLoadingIndicator()
         questionFactory?.requestNextQuestion()
@@ -49,12 +49,12 @@ final class MovieQuizPresenter: QuestionFactoryDelegate {
     }
     
     func didFailToLoadImage(with error: Error) {
-         DispatchQueue.main.async { [weak self] in
-             guard let self = self else { return }
-             let message = error.localizedDescription
-             self.viewController?.showNetworkError(message: message)
-         }
-     }
+        DispatchQueue.main.async { [weak self] in
+            guard let self = self else { return }
+            let message = error.localizedDescription
+            self.viewController?.showNetworkError(message: message)
+        }
+    }
     
     // MARK: - Public Methods
     
@@ -70,12 +70,13 @@ final class MovieQuizPresenter: QuestionFactoryDelegate {
     func switchToNextQuestion() {
         currentQuestionIndex += 1
     }
-
+    
     func convert(model: QuizQuestion) -> QuizStepViewModel {
         return QuizStepViewModel(
             image: UIImage(data: model.image) ?? UIImage(),
             question: model.text,
-            questionNumber: "\(currentQuestionIndex + 1)/\(questionsAmount)")
+            questionNumber: "\(currentQuestionIndex + 1)/\(questionsAmount)"
+        )
     }
     
     private func setButtonsEnabled(_ isEnabled: Bool) {
@@ -89,7 +90,7 @@ final class MovieQuizPresenter: QuestionFactoryDelegate {
     func noButtonClicked() {
         didAnswer(isYes: false)
     }
-
+    
     func didAnswer(isYes: Bool) {
         guard let currentQuestion = currentQuestion else { return }
         let isCorrectAnswer = isYes == currentQuestion.correctAnswer
@@ -100,13 +101,13 @@ final class MovieQuizPresenter: QuestionFactoryDelegate {
         
         proceedWithAnswer(isCorrect: isCorrectAnswer)
     }
-
+    
     private func proceedToNextQuestionOrResults() {
-        if self.isLastQuestion() {
-            let text = correctAnswers == self.questionsAmount ?
+        if isLastQuestion() {
+            let text = correctAnswers == questionsAmount ?
             "Поздравляем, вы ответили на 10 из 10!" :
             "Вы ответили на \(correctAnswers) из 10, попробуйте ещё раз!"
-
+            
             let viewModel = QuizResultsViewModel(
                 title: "Этот раунд окончен!",
                 text: text,
@@ -115,7 +116,7 @@ final class MovieQuizPresenter: QuestionFactoryDelegate {
             correctAnswers = 0
             currentQuestionIndex = 0
         } else {
-            self.switchToNextQuestion()
+            switchToNextQuestion()
             questionFactory?.requestNextQuestion()
             setButtonsEnabled(true)
         }
@@ -125,31 +126,31 @@ final class MovieQuizPresenter: QuestionFactoryDelegate {
         guard let statisticService = statisticService else {
             return "Ошибка: статистика недоступна"
         }
-
+        
         statisticService.store(correct: correctAnswers, total: questionsAmount)
         guard let bestGame = statisticService.bestGame else {
             return "Ошибка: лучшая игра недоступна"
         }
-
+        
         let totalPlaysCountLine = "Количество сыгранных квизов: \(statisticService.gamesCount)"
         let currentGameResultLine = "Ваш результат: \(correctAnswers)/\(questionsAmount)"
         let bestGameInfoLine = "Рекорд: \(bestGame.correct)/\(bestGame.total) (\(bestGame.date.dateTimeString))"
         let averageAccuracyLine = "Средняя точность: \(String(format: "%.2f", statisticService.totalAccuracy))%"
-
+        
         let resultMessage = [
             currentGameResultLine,
             totalPlaysCountLine,
             bestGameInfoLine,
             averageAccuracyLine
         ].joined(separator: "\n")
-
+        
         return resultMessage
     }
-
+    
     func proceedWithAnswer(isCorrect: Bool) {
-
+        
         viewController?.highlightImageBorder(isCorrectAnswer: isCorrect)
-
+        
         DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) { [weak self] in
             guard let self = self else { return }
             self.proceedToNextQuestionOrResults()
