+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -3,24 +3,24 @@ import UIKit
 // MARK: - MovieQuizViewController
 
 final class MovieQuizViewController: UIViewController, MovieQuizViewControllerProtocol {
-
+    
     // MARK: - IBOutlet
-
+    
     @IBOutlet private weak var noButton: UIButton!
     @IBOutlet private weak var yesButton: UIButton!
     @IBOutlet private weak var imageView: UIImageView!
     @IBOutlet private weak var textLabel: UILabel!
     @IBOutlet private weak var counterLabel: UILabel!
     @IBOutlet private weak var activityIndicator: UIActivityIndicatorView!
-
+    
     // MARK: - Private variables
     
     private var presenter: MovieQuizPresenter!
     private var alertPresenter: AlertPresenter!
     private var isButtonsEnabled = true
-
+    
     // MARK: - Lifecycle
-
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         presenter = MovieQuizPresenter(viewController: self)
@@ -29,13 +29,13 @@ final class MovieQuizViewController: UIViewController, MovieQuizViewControllerPr
         activityIndicator.hidesWhenStopped = true
         activityIndicator.startAnimating()
     }
-
+    
     override var preferredStatusBarStyle: UIStatusBarStyle {
         return .lightContent
     }
     
     // MARK: - Public Methods
-
+    
     func showCurrentQuestion(step: QuizStepViewModel) {
         imageView.image = step.image
         textLabel.text = step.question
@@ -44,30 +44,30 @@ final class MovieQuizViewController: UIViewController, MovieQuizViewControllerPr
         imageView.layer.borderWidth = 0
         imageView.layer.borderColor = UIColor.clear.cgColor
     }
-
+    
     func showQuizResults(result: QuizResultsViewModel) {
         let message = presenter.makeResultMessage()
-
+        
         let alertModel = AlertModel(
             title: result.title,
             message: message,
             buttonText: result.buttonText) { [weak self] in
-               self?.presenter.restartGame()
-           }
-       alertPresenter?.show(alertModel: alertModel)
-   }
-
+                self?.presenter.restartGame()
+            }
+        alertPresenter?.show(alertModel: alertModel)
+    }
+    
     func highlightImageBorder(isCorrectAnswer: Bool) {
         imageView.layer.masksToBounds = true
         imageView.layer.borderWidth = 8
         imageView.layer.borderColor = isCorrectAnswer ? UIColor.ypGreen.cgColor : UIColor.ypRed.cgColor
     }
-
+    
     func showLoadingIndicator() {
         activityIndicator.isHidden = false
         activityIndicator.startAnimating()
     }
-
+    
     func hideLoadingIndicator() {
         activityIndicator.isHidden = true
     }
@@ -95,15 +95,15 @@ final class MovieQuizViewController: UIViewController, MovieQuizViewControllerPr
     func showNetworkError(message: String) {
         hideLoadingIndicator()
         let alertModel = AlertModel(title: "Ошибка",
-                               message: message,
-                               buttonText: "Попробовать еще раз") { [weak self] in
+                                    message: message,
+                                    buttonText: "Попробовать еще раз") { [weak self] in
             guard let self = self else { return }
-
+            
             self.presenter.restartGame()
             self.presenter.switchToNextQuestion()
             self.showLoadingIndicator()
         }
-
+        
         alertPresenter?.show(alertModel: alertModel)
     }
