+++ MovieQuiz/Services/MoviesLoader.swift
@@ -12,24 +12,24 @@ protocol MoviesLoading {
 }
 
 struct MoviesLoader: MoviesLoading {
-
+    
     // MARK: - NetworkClient
-
+    
     private let networkClient: NetworkRouting
     
     init(networkClient: NetworkRouting = NetworkClient()) {
         self.networkClient = networkClient
-        }
-
+    }
+    
     // MARK: - URL
-
+    
     private var mostPopularMoviesUrl: URL {
         guard let url = URL(string: "https://imdb-api.com/en/API/Top250Movies/k_zcuw1ytf") else {
             preconditionFailure("Unable to construct mostPopularMoviesUrl")
         }
         return url
     }
-
+    
     func loadMovies(handler: @escaping (Result<MostPopularMovies, Error>) -> Void) {
         networkClient.fetch(url: mostPopularMoviesUrl) { result in
             switch result {
