+++ MovieQuiz/Services/StatisticService.swift
@@ -8,16 +8,16 @@ protocol StatisticService {
 }
 
 final class StatisticServiceImpl {
-
+    
     private enum Keys: String {
         case correct, total, bestGame, gamesCount
     }
-
+    
     private let decoder: JSONDecoder
     private let encoder: JSONEncoder
     private var userDefaults = UserDefaults.standard
     private let dateProvider: () -> Date
-
+    
     init(userDefaults: UserDefaults = .standard,
          decoder: JSONDecoder = JSONDecoder(),
          encoder: JSONEncoder = JSONEncoder(),
@@ -31,7 +31,7 @@ final class StatisticServiceImpl {
 }
 
 extension StatisticServiceImpl: StatisticService {
-
+    
     var gamesCount: Int {
         get {
             userDefaults.integer(forKey: Keys.gamesCount.rawValue)
@@ -40,7 +40,7 @@ extension StatisticServiceImpl: StatisticService {
             userDefaults.set(newValue, forKey: Keys.gamesCount.rawValue)
         }
     }
-
+    
     var correct: Int {
         get {
             userDefaults.integer(forKey: Keys.correct.rawValue)
@@ -49,7 +49,7 @@ extension StatisticServiceImpl: StatisticService {
             userDefaults.set(newValue, forKey: Keys.correct.rawValue)
         }
     }
-
+    
     var total: Int {
         get {
             userDefaults.integer(forKey: Keys.total.rawValue)
@@ -58,11 +58,11 @@ extension StatisticServiceImpl: StatisticService {
             userDefaults.set(newValue, forKey: Keys.total.rawValue)
         }
     }
-
+    
     var totalAccuracy: Double {
-       Double(correct) / Double(total) * 100
+        Double(correct) / Double(total) * 100
     }
-
+    
     var bestGame: BestGame? {
         get {
             guard
@@ -72,21 +72,21 @@ extension StatisticServiceImpl: StatisticService {
             }
             return bestGame
         }
-
+        
         set {
             let data = try? encoder.encode(newValue)
             userDefaults.set(data, forKey: Keys.bestGame.rawValue)
         }
     }
-
+    
     func store(correct: Int, total: Int) {
         self.correct += correct
         self.total += total
         self.gamesCount += 1
-
+        
         let date = dateProvider()
         let currentBestGame = BestGame(correct: correct, total: total, date: date)
-
+        
         if let previousBestGame = bestGame {
             if currentBestGame > previousBestGame {
                 bestGame = currentBestGame
