+++ MovieQuizPresenterTests/MovieQuizPresenterTests.swift
@@ -8,37 +8,6 @@
 import XCTest
 @testable import MovieQuiz
 
-class MovieQuizViewControllerMock: MovieQuizViewControllerProtocol {
-
-    func showCurrentQuestion(step: QuizStepViewModel) {
-    
-    }
-    
-    func showQuizResults(result: QuizResultsViewModel) {
-    
-    }
-    
-    func highlightImageBorder(isCorrectAnswer: Bool) {
-    
-    }
-    
-    func showLoadingIndicator() {
-    
-    }
-    
-    func hideLoadingIndicator() {
-    
-    }
-    
-    func setButtonsEnabled(_ isEnabled: Bool) {
-
-    }
-    
-    func showNetworkError(message: String) {
-    
-    }
-}
-
 final class MovieQuizPresenterTests: XCTestCase {
     func testPresenterConvertModel() throws {
         let viewControllerMock = MovieQuizViewControllerMock()
